Reorganize floating-point demos into separate directories
- Move each demo into its own subdirectory (arrayfsum, quad,
  fpadd, fpprint, fp_representation, arrayfsum_inline)
- Create individual Makefiles for each demo
- Update README.md to document new directory structure
- Create master Makefile in root to build all demos
- Fix DOS line endings in all source files
- Fix code style issues per checkpatch
- Add .gitignore to exclude build artifacts and .vscode/
