--- curs/chap-11-floating-point/01-all-demos/.gitignore
@@ -1,8 +1,0 @@
-/fp-m32
-/fp-m32-O2
-/fp-m32.s
-/fp-m32-O2.s
-/arrayfsumc
-/quadc
-/fpadd
-/fpprint
