+++ curs/chap-11-floating-point/arrayfsum/arrayfsuma.h
@@ -1,0 +1,8 @@
+/* SPDX-License-Identifier: BSD-3-Clause */
+
+#ifndef ARRAYFSUMA_H
+#define ARRAYFSUMA_H
+
+double array_fsum(double *value, int size);
+
+#endif
