+++ curs/chap-11-floating-point/arrayfsum_inline/Makefile
@@ -1,0 +1,14 @@
+CC = cc
+STACK_FLAGS = -Wl,-z,noexecstack
+BINARY = arrayfsum_inline.o
+
+all: $(BINARY)
+
+%.o : %.c
+	$(CC) -m64 -g -c -o $@ $<
+
+clean:
+	rm -f *.o
+	rm -f *~
+
+.PHONY: all clean
