+++ curs/chap-11-floating-point/quad/quada.h
@@ -1,0 +1,9 @@
+/* SPDX-License-Identifier: BSD-3-Clause */
+
+#ifndef QUADA_H
+#define QUADA_H
+
+int quad_roots(double a, double b, double c,
+	       double *root1, double *root2);
+
+#endif
